API config complete

+++ LivestockManagementSystem/app/Http/Controllers/PostController.php
@@ -4,11 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
-use App\Http\Requests\StorePostRequest;
-use App\Http\Requests\UpdatePostRequest;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -22,13 +30,13 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        dd('test');
+        //dd('test');
         $fields = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
         ]);
 
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
 
         return $post;
     }
@@ -46,6 +54,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modify', $post); //Authorization policy
+
         $fields = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -61,6 +71,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post); //Authorization policy
         $post->delete();
 
         return ['message' => 'The post was deleted'];
